requirements.txt
awslambdaric==2.0.4
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
PyYAML==6.0
requests==2.28.2
simplejson==3.17.2
urllib3==1.26.14
